+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 0.2.0 (WIP)
+- Addition: Web component API `<canvas is="paint-picture></canvas>`.
 - Breaking: Replace the colour type with `Colour` from gleam_community_colour.
 - Breaking: Replace `color_rgb` with `colour_rgb` and `colour_hex`.
 - Breaking: `display_on_canvas` and `interact_on_canvas` now take any css selector instead of a id.